"authentication credential" to replace "any means"
There is much criticism of line 24 because "any means" would be too vague to be reasonable, potentially criminalizing broad areas of useful research. Revision attempts to make it possible for security researchers to perform essential assessments and studies of weakness to better-inform defenders and improve security. Debate would inevitably be about what constitutes an authentication credential and controlled access, which also is helpful for security research and regulators.